netflixuax/requirements.txt
djangorestframework-simplejwt
python-decouple
requests
django-bootstrap4